Add gson ProtoTypeAdapter to third_party with fixes
Since the original version doesn't work with protov3 or protobuf lite,
which isn't particularly good since protov3 is supposed to be not only
the latest but also the greatest.

Hacked in a fix by making the template argument more generic (now it's
an AbstractMessage instead of a GeneratedMessage). This probably breaks
things in ways I don't yet understand.

Let's see.

Full details in NOTES.txt.