Replace **Model classes with protobufs for now
And load them in with a scoreboard.DomjudgeRest class that wraps up
a gson reader in the appropriate annotation magic and uses it to
translate fetched json into structured objects.

This takes a fair bit of the cruft out of the Executive class (the
commandline scoreboard viewer). Which is good because it's supposed to
essentially be a commandline parser that prints out its results in a
reasonably nice format.