Move the FakeModel into its own class for now
To make refactoring some of the other stuff easier without polluting
the diff with some WIP nonsense that shouldn't have been there in the
first place.

This also drops the number of imports in the main class and makes it
start to look a bit more like a main class, ie. no extraneous bits of
oddball logic.