Create a ScoreboardView; use in new resolver fxml
Code for the main class & the scoreboard view class are both much
easier to read now.

So we have (ScoreboardModel) -> (Executive) -> (ScoreboardView)

With the Executive just being in charge of handling input, routing it
through to ScoreboardModel, and telling ScoreboardView when it needs
to re-render, which is all it ever should be doing.

Now to make it animate we'll need some diffing. Or possibly not,
because all interesting changes affect one item at a time. The
exception is rejudgings, but they're not any more interesting as an
atomic op than a non-atomic one, so it's not worth all the
architectural scaffolding to support them as one.