Tool to serialise the entire api/v3 for a contest
The saved format is an EntireContest proto which has very little
structured information and is just bundling up all the responses from
api/v3 endpoints.

Makes replaying a contest offline later possible as long as there's
something able to read the pb.txt out again and marshal it back into
a contest --- which is the next step for this.