Step one toward using judgement_types for scoring
Added a mapping from judging outcome to judgement type. This mapping
doesn't seem to be presented properly by the API so for the moment this
is guessed using a DefaultJudgementTypes class.

Also added a User which is useful because of its being able to hold a
list of User.Role[] which should be printed out to the requester in a
warning message if they try to access something they're not allowed to
see.

Related to this, some of the endpoints necessary to download the whole
contest state are behind User.Role.jury or above, so all of the HTTP
request logic changed to use Optional<> instead and say something
diagnostic whenever a request fails. API call methods are annotated with
@RequiresRole(allOf / anyOf / any) to help a given requester figure out
what they're allowed to download.