Tidy up command line tool to abstract away API
Now all the tools operate on an EntireContest proto instead of knowing
anything specific about where the data is coming from.

This means a lot of code can be moved out of the entry point and into
one class that handles all the background fetching and updating
(although everything is still done synchronously for now).

Here's a straightforward way of checking nothing changed:

  * bazel run . -- --url https://www.domjudge.org/demoweb/api/ download | tee tmp.pb.txt | md5sum
  * bazel run . -- --file `pwd`/tmp.pb.txt download | md5sum