Create a ResolverController and scoreboard target
The command-line program has another option now: "resolver". This will
run a ResolverController on the scoreboard and repeatedly print out its
contents after each update.

For contests without judging data --- for example, the domjudge demoweb
site --- verdicts are guessed based on the final state of the scoreboard
to give a plausible set of judgings with the same resolution as the real
one.

This is the lst piece that was needed before being able to run a proper
resolver on the Android client instead of simply simulating the contest
in chronological order.