Implement support for CLICS api
Involved a large number of other changes to make it work. The API is
not as abstract as it was before but over time it should evolve back
into something that doesn't rely on a specific encoding of the data for
internal behaviour.

Of special note: replaced the basic scoreboard rank updater with one
based on a SplayList (really more of a SplaySet due to not designing the
interface very well, which is another thing to fix) so that rank updates
are O(logN). This makes the initial phase of preparing the resolver faster
for really large contests like NWERC 2017, since it involves submitting
and re-ranking ~1250 times.