Continue campaign of tidying up Android code + UI
Now with added CompletableFuture<>. Also now with thoroughly removed
android_library because I'm convinced I wasn't using it right and the
reason it fixed builds in the past was because desugaring support in
bazel wasn't quite there yet.

It's still not quite there but it's getting closer. Only two
catastrophic annotation crashes so far this week. Progress!