Beginning of a python script to replay contests
This just dumps out the entire contest in its final state at the moment
without applying any changes. The timestamp and duration elements are
hacked into the right shape by messing around with some internal
structures from the protobuf json formatter.

Once it's working it should be handy for testing live scoreboards with
running contests, since that has to work quite differently to the
resolver which only has to download the final state once and then it's
good to go.

An event feed might be necessary for this since it's the only kind of
long-polling the official API seems to have, and just polling the
scoreboard endpoint every 20 seconds isn't very nice (visually or
network-ily)

At the moment it is possible to set DOMJUDGE_CONTEST_API_IN_ROOT in
ClicsProto to true and download the contest proto, which comes out
identical to  the one fed into this script.