Step #1 towards making URL easy to enter correctly
By auto-detecting based on a set of common patterns from typical
installations of domjudge at various times. For the moment, any api
matching the following pattern should be picked up:

  https?://contest.example.com/(domjudge/)?(api/(v3/)?)?

For example running the console tool with...
  run scoreboards/console --url https://domjudge.bath.ac.uk resolver

...now works whereas before the API version and exact url (ending in
domjudge/api or domjudge/api/v3 for clics or domjudge3 respectively)
would have been needed too. The official format is preferred if there's
a choice, and detected by seeing if the API returns anything for
/groups, or /categories, or both.

Ugly details are returned in a ContestConfig.Source protobuf which will
be saved to disk in future and used for resuming a scoreboard later,
but at the moment all the data is unceremoniously pulled out straight
away and not checked again.