Add an Android activity to save contests offline
These can then be replayed later in resolver mode by clicking on them.
Authentication is not yet supported in the UI, and live scoreboards
(contemporaneous to a real contest) are not yet supported at all. Both
of these are next on the roadmap, followed by visual updates to make
this look less-strange on common form factors.