Have the console app show current row clearly
Previously it just showed the select problem, but now we'll show an
underline (OK actually a string of underscores) to indicate where the
resolver is at the moment.

This arguably makes the resolution more ugly but it looks more like a
real resolver, which helps to debug timing issues quickly without first
getting to a device with an emulator installed.