First steps at parsing the event-feed endpoint
Can eventually be paired with the hacked "event_id" field from the
scoreboard endpoint to jump into the right position in a contest. Not
yet sure that this is necessarily the best idea for a live scoreboard
specifically -- in this case it might be safest to just re-pull the
/scoreboard endpoint whenever something interesting happens in the
event feed, to avoid getting confused if eg. the active contest changes.