Proper implementation of CLICS' event feed.
Not supported for DOMjudgeV3, and probably never will be because the
event feed for that is in a totally different format and for this
minority of servers running old versions it will be easier to just
poll the endpoints in a loop for updates.

Call ContestDownloader.eventFeed(Contest) to see it in action, or to
save some time try running "scoreboards/console -- events" with the
url and username/password set appropriately.