Support optionals in DomJudgeV3 with wrappers
Tested on a variety of sites. Hopefully addresses #2.

The only difference I found when test-importing contests is that now the
Contest struct (correctly) doesn't include a scoreboard_freeze_duration
if a freeze_time is not set in the original DomJudgeV3 contest.

This means that scoreboards without a freeze will just go straight to
the final scoreboard for resolution, as they should.