Fix resolver's handling of evergreen contests.
We need to notify about submissions that happened when the freeze was
not applied, rather than the slightly-different "before the freeze"
(because this implies that if there is no freeze we shouldn't show
any submissions at all).

This is more bug than misunderstanding though. It didn't make any sense
like this, just most contests do have a freeze set.