Add an integration test for resolving NWERC 2007.
This was an interesting one because in old ICPC contests compile error
was treated as a penalty judgement instead of non-penalty. The demoweb
server hosting this doesn't seem to admit this in the API (but somehow
does know enough to calculate the scoreboard like that).

So to make the test pass we have to override the API and invent our own
version of the rules for this one small corner of the world.