Fix ugly font rendering bug (bad texture binding)
Code was assuming there's only one texture for all font variants and it
is bound 100% of the time. This isn't true. As soon as we create another
font or bind any different texture at all, things start to look strange.
Now fixed, hence possible to read text on a 1366x768 screen not just 4K.
This is the future.