Update command lines for new DOMjudge(Bug #4)
The commnands in the README directly reference the domjudge demoweb
site as a good place to test commands on if a real contest isn't
available.

As of April 2019 DOMjudge is using NWERC 2018 as an example contest
instead of NWERC 2007 so we need to change the group of participants
from "Participants" to "NWERC - Eindhoven University of Technology".
Setting just "14550" would also be fine but less transparent as to
what it means, so we don't do that.

Also: the scoreboards/resolver command had a bad command, the last
line "-- resolver" shouldn't be included. The program only has one
run mode so it doesn't parse commands like the console program does.