Pass "--contest {id}" to resolve one in particular
Eg. the following will resolve different contests:
  --url https://nwerc-judge.gehack.nl --contest practice
  --url https://nwerc-judge.gehack.nl --contest nwerc18

Default behaviour is to pick the first one in the array. If a contest
ID is given that is nonexistent on the server, there will be no attempt
to fall back to an existing one.

If the contest ID is not known, it can be left out. Future work: give
a nice console menu or GUI to choose what contest to resolve if it is
ambiguous.