Remove 'Teams' from ScoreboardModel's interfaces
Now callers have to fetch scoreboard.getTeamsModel() to call the
functions inside Teams. This is a little too clunky for the longer
term but will help with optimisation in the shorter term as we
can use a mix of static and dynamic implementations for the resolver,
which is the slowest class at the moment due to pre-simulating the
contest up to the freeze.