RanklistMutable: delete unused var for 2x speedup
getRank() can trigger a series of rotations in the SplayList, which is
expensive. This could be optimized, but it also could have been left out
in the first place as oldRank() was never used (the rank will be updated
implicitly by reinserting into the list if necessary.)

Performance improvement was verified using //javatests:contest_resolver.
Time to create the scoreboard for resolving went from 2s to 0.75s.