Resolver: Only judge submissions once all are in.
Insert all submissions, then run all judgements. Previously submissions
were judged immediately after inserting.

This had increased the number of resolution events in the NWERC 2017
resolution from 400 to 401. The reason for this was that one team solved
a problem before the freeze, but then made an attempt on it afterward
anyway just for fun.

After reworking the code to first judge all prefreeze submissions, and
only then add postfreeze judgements, this is back to 400 and none of the
tests need to be changed.