Interop: round seconds and nanos towards zero
The reason for doing this is that seconds must have the same sign as
nanoseconds, implying seconds should be rounded towards zero and
nanoseconds should represent whatever is left over in the direction
away from zero.

This was tested on the Eindhoven instance running DOMjudge 5.1.2, which
uses Domjudge REST API v2. This was released on 2016-06-11.