Console app: present a simple choice of contest
In the case that there are multiple contests available, we give the user
a chance to pick one. Otherwise we pick the only one available (or exit
with an error if there isn't even one contest available).

This should be folded back into the ContestDownloader class, or some
extension of it, as a callback which is applied once and then cached
thereafter.