Break up //contest into submodules
To smooth the transition, there's still an empty :contest target which
only pulls in all of the contest submodules (except :data_structures)
and makes all of them visible by exporting them.

The reason for leaving :data_structures out is that generally users
shouldn't need to know about this implementation detail. In fact it
should live somewhere completly different from //contest (and when that
is done, it'll be worth experimenting with SplayList<> vs some kind of
SqrtDecompositionList).