First stab: switch fullscreen to correct monitor.
This is better than nothing, but the main problem as written in the TODO
comment is that assuming the centre of the window is inside its
favourite monitor for fullscreening is not correct, and will not
identify any monitor at all in some cases.

We need to get the coverage of each monitor and use that. The only
reason for committing this at all is that development has been slow
recently, and this will make picking up the work later on easier.