First implementation of fallback fonts.
I am not entirely happy with this because the immutable allCodePoints is
not used for everything any more while codePointsToGlyphs needs to be
regenerated on each font size initialisation for proper ordering. This
is dirty and not multithreading-safe.

The solution will be to extract out generation of codePointsToGlyphs and
run it in some kind of constructor.

Fix #8