Add support for CLICS API redirects + Basic Auth
In DOMjudge and other CLICS servers the /contests endpoint is publicly
available however in Kattis the /contests endpoint redirects to a
canonical URL of /contests/ and is also stuck behind HTTP Basic Auth
requiring an API key.

We'll need to pass the login parameters through to the ApiDetective
- meaning it will offer the authentication up to all candidate endpoints
- in addition to handling the case of an HTTP 308 Permanent Redirect
if and only if it's pointing to the exact URL of /contests/ - this is
not ideal but necessary to avoid the case of misconfigured servers that
respond with a redirect to some intranet login page for nonexistent
endpoints instead of putting up a polite 404 page for the ApiDetective
to pick up on.