Put in a hack to work around Kattis ISO-8601 quirk
See some information in https://en.wikipedia.org/wiki/ISO_8601 and in
the official CLICS spec for how it's supposed to look. According to
the CLICS spec, +hh instead of +hh:mm is acceptable, but :s.mmm is
right out.

This will need a workaround forever due to back compatibility with
archived Kattis contests but it would be nice for Kattis to make this
more consistent later on.